feat: change ProtectedComponent
